requirements.txt
absl-py==2.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
aiosqlite==0.21.0
alembic==1.18.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
appdirs==1.4.4
astunparse==1.6.3
attrs==25.4.0
backoff==2.2.1
bcrypt==5.0.0
beautifulsoup4==4.13.5
build==1.4.0
cachetools==7.0.1
certifi==2026.1.4
cffi==2.0.0
cfgv==3.5.0
charset-normalizer==3.4.4
chromadb==1.1.1
click==8.1.8
cloudpickle==3.1.2
contourpy==1.3.3
crewai==1.9.3
crewai-tools==1.9.3
cryptography==46.0.4
cycler==0.12.1
decorator==5.2.1
defusedxml==0.7.1
deprecation==2.1.0
diskcache==5.6.3
distlib==0.4.0
distro==1.9.0
dm-tree==0.1.9
dnspython==2.8.0
docker==7.1.0
docstring_parser==0.17.0
durationpy==0.10
easydict==1.10
email-validator==2.3.0
et_xmlfile==2.0.0
fastapi==0.128.6
fastuuid==0.14.0
filelock==3.20.3
flatbuffers==25.12.19
fonttools==4.61.1
frozenlist==1.8.0
fsspec==2026.2.0
gast==0.4.0
google-api-core==2.30.0
google-auth==2.48.0
google-auth-oauthlib==1.0.0
google-cloud-core==2.5.0
google-cloud-storage==3.9.0
google-crc32c==1.8.0
google-pasta==0.2.0
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
greenlet==3.3.2
groq==0.37.1
grpcio==1.78.0
h11==0.16.0
h5py==3.15.1
hf-xet==1.2.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==0.36.2
identify==2.6.16
idna==3.11
importlib_metadata==8.7.1
importlib_resources==6.5.2
iniconfig==2.3.0
instructor==1.12.0
jax==0.4.30
jaxlib==0.4.30
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.3
json5==0.10.0
json_repair==0.25.3
jsonpatch==1.33
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
keras==2.15.0
kiwisolver==1.4.9
kubernetes==35.0.0
lancedb==0.5.7
langchain-core==1.2.9
langchain-groq==1.1.2
langsmith==0.7.0
libclang==18.1.1
litellm==1.81.9
lxml==6.0.2
Mako==1.3.10
Markdown==3.10.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mcp==1.23.3
mdurl==0.1.2
ml-dtypes==0.3.2
mmh3==5.2.0
mpmath==1.3.0
multidict==6.7.1
nodeenv==1.10.0
numpy==1.26.4
oauthlib==3.3.1
onnxruntime==1.24.1
openai==1.83.0
openpyxl==3.1.5
opentelemetry-api==1.34.1
opentelemetry-exporter-otlp-proto-common==1.34.1
opentelemetry-exporter-otlp-proto-grpc==1.34.1
opentelemetry-exporter-otlp-proto-http==1.34.1
opentelemetry-proto==1.34.1
opentelemetry-sdk==1.34.1
opentelemetry-semantic-conventions==0.55b1
opt_einsum==3.4.0
orjson==3.11.7
overrides==7.7.0
packaging>=23.1
pandas==1.5.3
passlib==1.7.4
pdfminer.six==20251230
pdfplumber==0.11.9
pillow==12.1.0
platformdirs==4.5.1
pluggy==1.6.0
pmlb==1.0.1.post3
portalocker==2.7.0
posthog==5.4.0
pre_commit==4.5.1
propcache==0.4.1
proto-plus==1.27.1
protobuf==5.29.6
psycopg2-binary==2.9.11
py==1.11.0
pyarrow==23.0.0
pyasn1==0.6.2
pyasn1_modules==0.4.2
pybase64==1.4.3
pycparser==3.0
pydantic==2.11.10
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.11.0
pylance==0.9.18
PyMuPDF==1.26.7
pyparsing==3.3.2
pypdfium2==5.4.0
PyPika==0.51.1
pyproject_hooks==1.2.0
pytest==9.0.2
pytest-asyncio==1.3.0
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
python-json-logger==4.0.0
python-multipart==0.0.22
pytube==15.0.0
pytz==2025.2
PyYAML==6.0.3
ratelimiter==1.2.0.post0
referencing==0.37.0
regex==2024.9.11
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
retry==0.9.2
rich==14.3.2
rpds-py==0.30.0
rsa==4.9.1
scikit-learn==1.3.2
scipy==1.15.3
semver==3.0.4
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.3
SQLAlchemy==2.0.46
sse-starlette==3.2.0
starlette==0.52.1
sympy==1.14.0
tenacity==9.1.4
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow==2.15.1
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.37.1
tensorflow-probability==0.23.0
termcolor==3.3.0
threadpoolctl==3.6.0
tiktoken==0.8.0
tokenizers==0.20.3
tomli==2.0.2
tomli_w==1.1.0
tqdm==4.67.3
typeguard==4.2.1
typer==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uuid_utils==0.14.0
uv==0.9.30
uvicorn==0.40.0
uvloop==0.22.1
virtualenv==20.36.1
watchfiles==1.1.1
websocket-client==1.9.0
websockets==16.0
Werkzeug==3.1.5
wrapt==1.14.2
xxhash==3.6.0
yarl==1.22.0
ydata-synthetic==1.4.0
youtube-transcript-api==1.2.4
zipp==3.23.0
zstandard==0.25.0
